Fix code issue related to phosphorus bump
https://docs.opendaylight.org/en/latest/release-notes/upgrade-process.html

+++ lighty-examples/lighty-controller-springboot-netconf/src/main/java/io/lighty/core/controller/springboot/rest/NetconfDeviceRestService.java
@@ -41,6 +41,7 @@ import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.topology.NodeBuilder;
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.topology.NodeKey;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint16;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -138,7 +139,7 @@ public class NetconfDeviceRestService {
             .setNodeId(nodeId)
             .addAugmentation(new NetconfNodeBuilder()
                 .setHost(new Host(new IpAddress(new Ipv4Address(deviceInfo.getAddress()))))
-                .setPort(new PortNumber(deviceInfo.getPort()))
+                .setPort(new PortNumber(Uint16.valueOf(deviceInfo.getPort())))
                 .setCredentials(new LoginPasswordBuilder()
                         .setUsername(deviceInfo.getUsername())
                         .setPassword(deviceInfo.getPassword())
